backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.26.2
pytesseract==0.3.10
scikit-learn==1.3.2
spacy==3.7.2
python-docx==1.0.1
Pillow==10.1.0
pandas==2.1.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
